change to initialize ActionMailer::Base.deliveries to an emtpy array
Not sure what's causing the behavior, but sometimes when we run the full test suite, the add_students feature spec fails because the deliveries is 4 instead of 2. Resetting the deliveries array to be empty makes the test consistent and gaurantees that there aren't extra things hanging around in the array.